Prevent flow navigation on clicking workflow status toggle (#1404)
Fixes OPS-2662.

## Additional Notes

This could be a regression of
https://github.com/openops-cloud/openops/pull/1242

Since we are wrapping each cell in a row with *Link*, adding event
handler directly to the tooltip is not working all the time. So we now
do not wrap the *Flow status toggle* with Link. This prevents the flow
navigation on clicking the status toggle.

I've added a new prop to DataTable component to control interactive
columns that should not be wrapped with *Link*