Force dynamic properties to be mandatory (#1469)
Fixes OPS-2861
The problem is that if the property is not mandatory, the LLM
(rightfully) doesn't pass it when building the workflow, which can lead
to bugs such as https://github.com/openops-cloud/openops/pull/1438.

In practice, dynamic properties will always return either a sub
property, or {}. So they can be marked as required=true.
This will serve as a guardrail for any future properties that will be
added.