Fix slack interactions by normalising emoji text (#1600)
Fixes OPS-2996.

## Additional Notes

Button text is being normalised by Slack in the response payload and
this lead to the conflict to resume the flow. Now we match the
normalised text response from slack with the normalised original action
text and then we return the original text in the updated message.

For multi select, we are already sending the userSelection.value which
slack doesn't normalise and need not be fixed.


## For future reference

- I've first added emoji: false prop to SlackText and although this
fixes the bug, Slack wouldn't render emojis in the UI when user uses
emoji in colon notation like ":emoji:"
- I've then tried normalisation in request action and then compare it
with original action in wait for interaction and then send the emojified
string back to the user.
- We've decided to send back the original text (either emoji or text
notation) to the user.
- So I've moved the normalisation to wait for interaction module and
return the original action to the user. For multi - select (Arrays), we
already return the value prop which slack doesn't normalise.
- As as alternate fix, I've tried
https://github.com/openops-cloud/openops/pull/1615, but this would
require user to add value prop for actions when using send message
action. So we ignore this method although it would've been the cleanest
approach.